Edit and Delete Articles functionality restricted to admins and authors only.

+++ src/main/java/blog/controllers/ArticleController.java
@@ -8,6 +8,7 @@ import blog.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.Banner;
 import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.authentication.AnonymousAuthenticationToken;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Controller;
@@ -57,6 +58,12 @@ public class ArticleController {
             return "redirect:/";
         }
 
+        if(!(SecurityContextHolder.getContext().getAuthentication() instanceof AnonymousAuthenticationToken)){
+            UserDetails principal = (UserDetails) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+            User entityUser = this.userRepository.findByEmail(principal.getUsername());
+            model.addAttribute("user", entityUser);
+        }
+
         Article article = this.articleRepository.findOne(id);
 
         model.addAttribute("article", article);
@@ -73,6 +80,10 @@ public class ArticleController {
         }
         Article article = this.articleRepository.findOne(id);
 
+        if(!isUserAuthorOrAdmin(article)){
+            return "redirect:/article/" + id;
+        }
+
         model.addAttribute("view", "article/edit");
         model.addAttribute("article", article);
 
@@ -86,6 +97,10 @@ public class ArticleController {
             return "redirect:/";
         }
         Article article = this.articleRepository.findOne(id);
+
+        if(!isUserAuthorOrAdmin(article)){
+            return "redirect:/article/" + id;
+        }
 
         article.setContent(articleBindingModel.getContent());
         article.setTitle(articleBindingModel.getTitle());
@@ -104,6 +119,10 @@ public class ArticleController {
 
         Article article = this.articleRepository.findOne(id);
 
+        if(!isUserAuthorOrAdmin(article)){
+            return "redirect:/article/" + id;
+        }
+
         model.addAttribute("article", article);
         model.addAttribute("view", "article/delete");
 
@@ -117,8 +136,20 @@ public class ArticleController {
             return "redirect:/";
         }
         Article article = this.articleRepository.findOne(id);
+
+        if(!isUserAuthorOrAdmin(article)){
+            return "redirect:/article/" + id;
+        }
         this.articleRepository.delete(article);
 
         return "redirect:/";
+    }
+
+    private boolean isUserAuthorOrAdmin(Article article){
+        UserDetails user = (UserDetails) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+        User userEntity = this.userRepository.findByEmail(user.getUsername());
+
+        return userEntity.isAdmin() || userEntity.isAuthor(article);
     }
 }
